Upgrade to Ruby 2.7.1
Ruby 2.2.3 was EOL'ed in March 2018. I get the following warning trying
to install it:

> WARNING: ruby-2.2.3 is past its end of life and is now unsupported.
> It no longer receives bug fixes or critical security updates.

On Ubuntu 20.04, I was not able to easily complete the installation of
2.2.3 with rbenv - it seemed to require older versions of development
libraries that might be hard to install on Ubuntu 20.04.

Jekyll's broadly compatible with most Ruby versions (and the site itself
is static once generated), so this isn't a huge deal, but the
`.ruby-version` file does ensure all developers get a consistent
experience. Using a more modern version of Ruby should remove a hurdle
for new contributors to get started. Ruby 2.7.1 is the latest stable
version, and the site does work as-is (locally) with this version of
Ruby - though I plan to also update some gems to more recent versions in
a follow-up commit.

I'd like to follow all this up with some other changes - perhaps
addressing #14 and #9 as a start. But it seemed like getting things up
to date and easier to work with was a good first step.