Update Gems
I ran `bundle update` to pull the latest versions of our gems. Most were
relatively minor upgrades. For example, we went from Jekyll 4.0 to 4.1.

In addition to this, I added version constraints to the Gemfile. For
example, `~> 4.1` means we'll take any Jekyll version `>= 4.1` but it
will prevent upgrading to `5.0` or anything higher (since that might
come with breaking changes we need to account for).