Add OSC authors to _config.yml
This commit is the first step for #9.

The idea is to add each author to `_config.yml` with a key for their
name. In the future, we can add additional (perhaps optional) keys for
GitHub, Twitter, email, etc. We can then update the post layout to use
whatever of these keys are available. When the author creates a post,
their info only needs to be in `_config.yml` once. As long as the post
references their author key, it can pull the info it needs.

To get this rolling, I've generated a "key"(username?) for each author
by downcasing their name and replacing spaces with `-`. These can be
adjusted for authors' preferences in the future, but I needed something
to start with. All authors have a `name` right now, but that's currently
the only field. (We should plan on supporting others.)

All this was done primarily with a script, which I've included in a new
`scripts` directory. (It seemed prudent to save stuff like this, though
it's also fine with me if we'd prefer to discard it.) The script makes
the changes to each post and spits out a yml file with authors. Then I
did one or two manual touch-ups and put the authors in `_config.yml`.

With this commit, the site remains in a working state, with no visible
changes. Although we've moved some "config" around, the posts themselves
still display the author name, exactly as they did before.