Fix IRC Mixed Content Warning
There's an error on the IRC page:

> Blocked loading mixed active content
> "http://webchat.freenode.net/?channels=%23%23catholic"

I'm fairly certain this is due to an `http` iframe on an `https` page.
(But this is difficult to test since I don't have https set up locally.)
Making the iframe https _should_ fix it. And while we're at it, the rest
of the links on the page can be https too.