Combine search and archive pages
We can streamline the navigation for the site by combining the Search
and Archive pages, putting the search bar at the top of the archive
page.

The search bar loads asynchronously with JS, so placing it in a
fixed-height div prevents the page content from jumping around when the
search bar loads.

We'll redirect the `/search/` and `/search/site/` (which used to
redirect to `/search/`) paths to `/archive/`.