Add CatholicDevs Slack Links
This commit changes the "IRC" page to a "Chat" page. We're still linking
to the `##catholic` channel on Freenode, but we're now also linking to
the CatholicDevs Slack workspace that John's been building.

In addition, I added a link to the Slack workspace as a social icon in
our footer.

There's no way to allow new users to join a Slack workspace without an
invitation. To work around this, we're using
https://communityinviter.com to provide automatic Slack invitations to
anyone who requests one from this site. I don't expect an overwhelming
volume here, but we can flip a switch to require approval of these
invitations at any time using Community Inviter.

Completes #30