Use post previews on the homepage
Currently, the homepage is a list of full-length posts, paginated with 7
per page. This requires a lot of scrolling, and it seems like a cleaner
design would be to show a preview of each post in the list on the
homepage, but require clicking into the post itself to continue reading.
The homepage looks a little cleaner, and you're more immersed in the
post itself (no other posts, and the comments are expanded) when you
read one.

There are probably a few different strategies for generating the
preview. The one I'm using here takes the first 40 words. It's somewhat
important to render HTML and markdown before truncating - otherwise
posts that have HTML or markdown content near the top might have
non-displayed words counting against them.