Improve post previews on homepage
In the previous commit, I added post previews to the homepage. Here, I'm
improving that experience. The first iteration of this functionality
stripped HTML and rendered the post preview as plain text. This was
required because it did simple text truncation, and it's problematic to
truncate in the middle of an HTML tag.

I'm now improving on that by using a plugin to do smart_truncate. This
plugin is aware of HTML markup and will ensure the HTML tag is closed
appropriately when we truncate. This means HTML will render correctly in
the preview even when we truncate! This is possible due to the plugin I
found referenced here:
https://yingtongli.me/blog/2016/07/12/html-truncate.html. We're allowed
to use it because it's open source under the GNU Affero GPL. Small
bummer that it's distributed as a ruby file rather than a gem and it
doesn't come with unit tests, but the code's short enough I don't view
this as a big concern.

I also double the number of words included in the preview to provide a
little more context to each post.