Migrate Deployment to GitHub Actions
Our current deployment uses the traditional GitHub Pages Jekyll build.
That is, GitHub Pages performs the build for us and deploys the result.
The GitHub Pages Jekyll build process is a little bit restrictive. We're
limited to Jekyll 3.9 and we can't use plugins that aren't in the
[approved list](https://pages.github.com/versions/).

A newer approach to deploying a Jekyll site with GitHub Pages uses
GitHub Actions. A GitHub Action is GitHub's own solution to generic
CI/CD. Therefore, our GitHub Action can build any version of Jekyll with
any plugins. The action pushes the result of this build to the
`gh-pages` branch, and then we deploy that branch as a pre-built static
site on GitHub Pages.

All of this is done following the
[guide](https://jekyllrb.com/docs/continuous-integration/github-actions/)
on the Jekyll website.

After this commit is merged, the Action will run (creating the branch),
but we won't be publishing the new branch 'til we manually edit the
repository settings to deploy from `gh-pages` instead of `master`. (This
gives us time to verify the build worked, etc.)