Fix broken link in about.md
`about.md` contained a link to `CONTRIBUTING.md`. This link actually
worked when viewing `about.md` as a markdown file - `CONTRIBUTING.md` is
in the same directory so a relative link works fine. However, when
Jekyll builds the site, `about.md` turns into an HTML page at
`/about/index.html`, and the relative link to `CONTRIBUTING.md` breaks.
(Also, I don't think we're even publishing CONTRIBUTING.md to the web --
which is fine, since people can read it as markdown on GitHub.)

The fix is simple -- just link to
https://github.com/opensourcecatholic/opensourcecatholic.github.io/blob/master/CONTRIBUTING.md