Replace smart_truncate with Gem
In d86e1a2, we introduced post previews on the homepage (so you get a
preview on the homepage, but have to visit the post itself to continue
reading). To do so, we added a Jekyll "plugin" which is actually just a
single ruby file -- `smart_truncate.rb`. (This plugin is open source and
came from https://yingtongli.me/git/smart_truncate/tree/smart_truncate.rb.)

That plugin works, but I was disappointed there wasn't an easier way to
do this (a built-in Jekyll feature or a plugin distributed as a gem).
So, I built my own
[jekyll_html_truncatewords](https://github.com/mkasberg/jekyll_html_trunctewords)
plugin. This plugin is distributed as a gem and has better unit test
coverage than the `smart_truncate.rb` file (which had no apparent unit
tests), so I feel better about using it. My intent is to make this a
reliable and well-maintained Jekyll plugin for use here and for anyone
else who wants to use it.

In this commit, I remove the `_plugins/smart_truncate.rb` file and
replace it with my `jekyll_html_truncatewords` gem. The functionality is
identical, and I spot-checked several rendered HTML pages which look
identical with either plugin. The end result is that we don't have to
maintain a copy of that `smart_truncate.rb` file and instead we can
update the gem like we would for any other Jekyll gem/plugin.