Fix destination file conflict
We publish posts in Jekyll based on a `<year>/<post-name>.html` scheme.
Ex. <https://www.opensourcecatholic.com/2010/social-media-policies.html>

This is fine, but it is resulting in conflicts for a couple of our
files:

```
Conflict: The following destination is shared by multiple files.
          The written file may end up with unexpected contents.
          ~/code/opensourcecatholic.github.io/_site/2010/social-media-policies.html
           - ~/code/opensourcecatholic.github.io/_posts/2010-06-21-social-media-policies.md
           - ~/code/opensourcecatholic.github.io/_posts/2010-07-08-social-media-policies.md
```

```
Conflict: The following destination is shared by multiple files.
          The written file may end up with unexpected contents.
          ~/code/opensourcecatholic.github.io/_site/2010/looking-for-custom-web-developer-with-extensive-wordpress-knowledge.html
           - ~/code/opensourcecatholic.github.io/_posts/2010-09-16-looking-for-custom-web-developer-with-extensive-wordpress-knowledge.md
           - ~/code/opensourcecatholic.github.io/_posts/2010-09-17-looking-for-custom-web-developer-with-extensive-wordpress-knowledge.md
```

There were only two conflicts. In the first case, I resolved the
conflict by appending a `-0` to one of the URLs to keep both posts. In
the second case, the posts were nearly identical anyway, so I chose to
keep the one with better formatting and delete the other.

This resolves all outstanding warnings we get when building the site.