Upgrade to Ruby 3.0
Ruby 3 should be faster than Ruby 2, which should mean a little less
time waiting for jekyll to build the site.

[Ruby 3.0.0 Released](https://www.ruby-lang.org/en/news/2020/12/25/ruby-3-0-0-released/)

Aside from the speed increase, it isn't a bad idea to keep up-to-date
with the latest version of Ruby.

Jekyll is compatible with Ruby 3 with the caveat that you need to add
WEBrick as a dependency when upgrading: https://github.com/jekyll/jekyll/issues/8523

I've run `jekyll serve` locally and the site builds without errors and
looks good in a browser. Especially considering this is a static site,
that's good enough for me!